Extract routes into modules
