+++ matchbox/src/server/routes/mod.rs
@@ -1,0 +1,4 @@
+pub mod error;
+pub mod sandbox;
+
+pub type ApiResult<T> = Result<T, error::ApiError>;
